Database.Sql.Hive.Parser

Documentation

statementParser :: Parser (HiveStatement RawNames Range)

emptyParserScope :: ParserScope

parse :: Text -> Either ParseError (HiveStatement RawNames Range)

parse consumes a statement, or fails

parseAll :: Text -> Either ParseError (HiveStatement RawNames Range)

parseAll consumes all input as a single statement, or fails

parseMany :: Text -> Either ParseError [HiveStatement RawNames Range]

parseMany consumes multiple statements, or fails

parseManyAll :: Text -> Either ParseError [HiveStatement RawNames Range]

parseManyAll consumes all input multiple statements, or fails

parseManyEithers :: Text -> Either ParseError [Either (Unparsed Range) (HiveStatement RawNames Range)]

parseManyEithers consumes all input as multiple (statements or failures) it should never fail

optionBool :: Parser a -> Parser Bool

statementP :: Parser (Statement Hive RawNames Range)

useP :: Parser (Use Range)

analyzeP :: Parser (Analyze RawNames Range)

insertDirectoryPrefixP :: Parser (Range, InsertDirectoryLocale Range, Location Range)

insertDirectoryP :: (QueryPrefix, InvertedFrom) -> Parser (InsertDirectory RawNames Range)

insertDirectoryLocaleP :: Parser (InsertDirectoryLocale Range)

insertDirectoryPathP :: Parser (Location Range)

staticPartitionSpecItemP :: Parser (StaticPartitionSpecItem RawNames Range)

staticPartitionSpecP :: Parser ([StaticPartitionSpecItem RawNames Range], Range)

type PartitionDecider = Either (StaticPartitionSpecItem RawNames Range) (DynamicPartitionSpecItem RawNames Range)

dynamicPartitionSpecItemP :: Parser (DynamicPartitionSpecItem RawNames Range)

partitionSpecDeciderP :: Parser PartitionDecider

partitionSpecP :: Parser ()

truncatePartitionStatementP :: Parser (TruncatePartition RawNames Range)

describeP :: Parser Range

showP :: Parser Range

createFunctionPrefixP :: Parser Range

createFunctionP :: Parser Range

dropFunctionPrefixP :: Parser Range

dropFunctionP :: Parser Range

alterTableSetLocationP :: Parser (AlterTableSetLocation RawNames Range)

alterTableSetTblPropertiesP :: Parser Range

alterPartitionP :: Parser (HiveStatement RawNames Range)

setP :: Parser (SetProperty Range)

reloadFunctionP :: Parser Range

insertBehaviorHelper :: InsertBehavior Range -> Maybe TablePartition -> InsertBehavior Range

insertP :: (QueryPrefix, InvertedFrom) -> Parser (Insert RawNames Range)

loadDataInPathP :: Parser (Insert RawNames Range)

deleteP :: Parser (Delete RawNames Range)

truncateP :: Parser (Truncate RawNames Range)

type QueryPrefix = Query RawNames Range -> Query RawNames Range

emptyPrefix :: QueryPrefix

withP :: Parser QueryPrefix

querySelectP :: (QueryPrefix, InvertedFrom) -> Parser (Query RawNames Range)

queryP :: (QueryPrefix, InvertedFrom) -> Parser (Query RawNames Range)

queryPHelper :: QueryPrefix -> InvertedFrom -> Bool -> Parser (Query RawNames Range)

distinctP :: Parser Distinct

explainP :: Parser (Statement Hive RawNames Range)

tableAliasP :: Parser (TableAlias Range)

columnAliasP :: Parser (ColumnAlias Range)

createSchemaPrefixP :: Parser Range

ifNotExistsP :: Parser (Maybe Range)

commentP :: Parser Range

locationP :: Parser (Location Range)

createSchemaP :: Parser (CreateSchema RawNames Range)

createViewPrefixP :: Parser Range

createViewP :: Parser (CreateView RawNames Range)

data CreateTablePrefix r a

Constructors

CreateTablePrefix

Fields

createTablePrefixInfo :: a

createTablePrefixPersistence :: Persistence a

createTablePrefixExternality :: Externality a

createTablePrefixIfNotExists :: Maybe a

createTablePrefixName :: CreateTableName r a

createTablePrefixP :: Parser (CreateTablePrefix RawNames Range)

createTableP :: Parser (CreateTable Hive RawNames Range)

createTableLikeP :: Parser (CreateTable Hive RawNames Range)

propertyP :: Parser (HiveMetadataProperty Range)

storedAsP :: Parser Range

createTableStandardP :: Parser (CreateTable Hive RawNames Range)

tblPropertiesP :: Parser (HiveMetadataProperties Range)

delimitedP :: Parser Range

ifExistsP :: Parser Range

dropTableP :: Parser (DropTable RawNames Range)

alterTableRenameTablePrefixP :: Parser (Range, TableName RawNames Range)

alterTableRenameTableP :: Parser (AlterTable RawNames Range)

alterTableRenameColumnPrefixP :: Parser (Range, TableName RawNames Range)

unqualifiedColumnNameP :: Parser (UQColumnName Range)

alterTableRenameColumnP :: Parser (AlterTable RawNames Range)

alterTableAddColumnsPrefixP :: Parser (Range, TableName RawNames Range)

alterTableAddColumnsP :: Parser (AlterTable RawNames Range)

grantP :: Parser (Grant Range)

revokeP :: Parser (Revoke Range)

integerP :: Parser (Int, Range)

countingSepBy1 :: (Integer -> Parser b) -> Parser c -> Parser [b]

introduceAliases :: Set Text -> ParserScope -> ParserScope

tablishToTableAlias :: Tablish RawNames Range -> Set Text

tableNameToTableAlias :: OQTableName Range -> Set Text

fromP :: Parser (SelectFrom RawNames Range)

type InvertedFrom = Maybe (SelectFrom RawNames Range)

noInversion :: InvertedFrom

invertedFromP :: Parser InvertedFrom

selectP :: InvertedFrom -> Parser (Select RawNames Range)

handlePositionalReferences :: Expr RawNames Range -> PositionOrExpr RawNames Range

selectGroupP :: Parser (SelectGroup RawNames Range)

selectClusterP :: Parser ()

selectClusterP parses for either clusterby or distributeby/sortby T478023 - implement clusterby in select datatype

qualifiedTableNameP :: Parser (Text, Text, Range, Range)

checkTableNameInScopeP :: Text -> Parser ()

The columnName parser has been overhauled with checks for table names. If a scope is present (i.e. in a select statement), the table name must be a member of the tableAlias list for the parser to succeed. Otherwise, the table parser fails and execution tries the next parser choice.

Should the scope not be set, e.g. when selectP is performing lookahead to build scope, this check is skipped.

selectStarP :: Parser (Selection RawNames Range)

tableNameP :: Parser (OQTableName Range)

arrayAccessP :: Parser (Expr RawNames Range -> Expr RawNames Range)

structFieldNameP :: Parser (StructFieldName Range)

structAccessP :: Parser (Expr RawNames Range -> Expr RawNames Range)

columnNameP :: Parser (OQColumnName Range)

selectionP :: Integer -> Parser (Selection RawNames Range)

makeColumnAlias :: Range -> Text -> Parser (ColumnAlias Range)

makeTableAlias :: Range -> Text -> Parser (TableAlias Range)

makeDummyAlias :: Range -> Integer -> Parser (ColumnAlias Range)

makeExprAlias :: Expr RawNames Range -> Integer -> Parser (ColumnAlias Range)

exprP :: Parser (Expr RawNames Range)

parenExprP :: Parser (Expr RawNames Range)

caseExprP :: Parser (Expr RawNames Range)

fieldTypeP :: Parser (Expr RawNames Range)

functionExprP :: Parser (Expr RawNames Range)

orderTopLevelP :: Parser (Range, [Order RawNames Range])

orderInWindowClauseP :: Parser [Order RawNames Range]

orderExprP :: Bool -> Bool -> Parser (Range, [Order RawNames Range])

directionP :: Parser (OrderDirection (Maybe Range))

nullsP :: Parser (NullPosition (Maybe Range))

frameP :: Parser (Frame Range)

frameBoundP :: Parser (FrameBound Range)

overP :: Parser (OverSubExpr RawNames Range)

windowExprP :: Range -> Parser (WindowExpr RawNames Range)

partialWindowExprP :: Range -> Parser (PartialWindowExpr RawNames Range)

windowNameP :: Parser (WindowName Range)

partitionP :: Parser (Partition RawNames Range)

dataTypeP :: Parser (DataType Range)

existsExprP :: Parser (Expr RawNames Range)

columnExprP :: Parser (Expr RawNames Range)

variableSubstitutionP :: Parser (Expr RawNames Range)

exprWithArrayOrStructAccessP :: Parser (Expr RawNames Range)

unOpP :: Text -> Parser (Expr RawNames Range -> Expr RawNames Range)

unaryPrefixExprP :: Parser (Expr RawNames Range)

notOperatorP :: Parser (Expr RawNames Range -> Expr RawNames Range)

unarySuffixExprP :: Parser (Expr RawNames Range)

binOpP :: Text -> Parser (Expr RawNames Range -> Expr RawNames Range -> Expr RawNames Range)

bitwiseXorExprP :: Parser (Expr RawNames Range)

productExprP :: Parser (Expr RawNames Range)

sumExprP :: Parser (Expr RawNames Range)

stringExprP :: Parser (Expr RawNames Range)

bitwiseAndExprP :: Parser (Expr RawNames Range)

bitwiseOrExprP :: Parser (Expr RawNames Range)

inExprP :: Parser (Expr RawNames Range)

betweenExprP :: Parser (Expr RawNames Range)

likeExprP :: Parser (Expr RawNames Range)

mkBinOp :: (Text, a) -> Expr r a -> Expr r a -> Expr r a

inequalityExprP :: Parser (Expr RawNames Range)

equalityExprP :: Parser (Expr RawNames Range)

notExprP :: Parser (Expr RawNames Range)

andExprP :: Parser (Expr RawNames Range)

orExprP :: Parser (Expr RawNames Range)

singleTableP :: Parser (Tablish RawNames Range)

singleTableWithViewsP :: Parser (Tablish RawNames Range)

lateralViewP :: Parser (Tablish RawNames Range -> Tablish RawNames Range)

optionalParensP :: Parser a -> Parser a

manyParensP :: Parser a -> Parser a

tablishP :: Parser (Tablish RawNames Range)

joinP :: Parser (Tablish RawNames Range -> Tablish RawNames Range)

outerJoinTypeP :: Parser (JoinType Range)

innerJoinTypeP :: Parser (JoinType Range)

crossJoinTypeP :: Parser (JoinType Range)

semiJoinTypeP :: Parser (JoinType Range)

constantP :: Parser (Constant Range)